requirements.txt
opencv-python
pytesseract
numpy
Pillow
ultralytics
selenium